Title: 특정 수 도달하기, Time: 133ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 260212/특정 수 도달하기/reaching-specific-number.java
@@ -2,11 +2,11 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         // Please write your code here.
-        Scanner sc = new Scanner(System.in);
+         Scanner sc = new Scanner(System.in);
 
         int[] arr = new int[10];
-        double sum = 0;
-        int cnt = 0;
+        int sum = 0;
+        double cnt = 0;
 
         for (int i = 0; i < arr.length; i++) {
             arr[i] = sc.nextInt();
@@ -18,6 +18,6 @@ public class Main {
             cnt++;
         }
 
-        System.out.print(sum + " " + (sum / cnt));
+        System.out.print(sum + " " + (double)(sum / cnt));
     }
 }
